A lot to describe
There was one final restructuring done to the whole app. First success was in solving how to get props to work and render my movie images and synopses. Next success was in adding more props into the movieDetails array and successfully passing those down. Lots and lots of time was spent trying to figure out how to get values from the lowest components and pass those values up through functions, not to mention tracking those functions being passed down as props from the top down. Still learning how and where to properly use state. Reviews currently seem to work and display/update in the list. In the process of moving the star component up to the top and properly displaying.