Project completed!
Accomplishments today: found a way to move the value from the Stars component all the way up to the MovieList, and to move the props from movieDetails of the starArray all the way down to the Stars component. Changed several functions to reflect the submit button changing the state for both Stars and the Review portion of the ReviewForm. Once this functionality was in place, I wanted to find a way to clear the values in the review form and the selection for the star value. These were handled with refs, and ultimately the handleChange function for the data in the reviewInputContent textarea was removed. Instead, the value in that box was passed up through the function by  replacing its previous state value as this.reviewTextArea.current.value. Lots and lots of research needed to accomplish this all, but the project was a success.